+++ packages/@expo/cli/src/start/platforms/android/__tests__/gradle-test.ts
@@ -47,7 +47,7 @@ describe(assembleAsync, () => {
       variant: 'something',
       buildCache: true,
       appName: 'foobar',
-      port: 19000,
+      port: 8081,
     });
 
     expect(spawnAsync).toHaveBeenCalledWith(
@@ -60,7 +60,7 @@ describe(assembleAsync, () => {
         'test',
         '--configure-on-demand',
         '--build-cache',
-        '-PreactNativeDevServerPort=19000',
+        '-PreactNativeDevServerPort=8081',
       ],
       { cwd: '/android', stdio: 'inherit' }
     );
@@ -81,11 +81,11 @@ describe(assembleAsync, () => {
 
 describe(installAsync, () => {
   it(`installs`, async () => {
-    await installAsync('/android', { variant: 'something', appName: 'foobar', port: 19000 });
+    await installAsync('/android', { variant: 'something', appName: 'foobar', port: 8081 });
 
     expect(spawnAsync).toHaveBeenCalledWith(
       '/android/gradlew',
-      ['foobar:installSomething', '-PreactNativeDevServerPort=19000'],
+      ['foobar:installSomething', '-PreactNativeDevServerPort=8081'],
       { cwd: '/android', stdio: 'inherit' }
     );
   });