+++ packages/expo-auth-session/src/AuthSession.ts
@@ -152,7 +152,7 @@ export function getRedirectUrl(path?: string): string {
  *   path: 'redirect'
  * });
  * // Development Build: my-scheme://redirect
- * // Expo Go: exp://127.0.0.1:19000/--/redirect
+ * // Expo Go: exp://127.0.0.1:8081/--/redirect
  * // Web dev: https://localhost:19006/redirect
  * // Web prod: https://yourwebsite.com/redirect
  *
@@ -162,7 +162,7 @@ export function getRedirectUrl(path?: string): string {
  *   isTripleSlashed: true,
  * });
  * // Development Build: scheme2:///
- * // Expo Go: exp://localhost:19000
+ * // Expo Go: exp://localhost:8081
  * // Web dev: https://localhost:19006
  * // Web prod: https://yourwebsite.com
  * ```